requirements.txt
openai>=1.42.0
pandas
faiss-cpu
numpy